Moved tests from main to login
